requirements.txt
# Sistema de Email Marketing Empresarial - Dependências
# Automated Lead Generator

# === CORE DEPENDENCIES ===
# Manipulação de dados
pandas>=2.0.0

# Configurações de ambiente
python-dotenv>=1.0.0

# Excel support
openpyxl>=3.1.0

# === TRACKING E ANALYTICS ===
# Servidor web para tracking
Flask>=2.3.0

# Banco de dados (já incluído no Python)
# sqlite3 - built-in

# === EMAIL E NETWORKING ===
# SMTP (já incluído no Python)
# smtplib - built-in
# email - built-in

# Requisições HTTP (para verificações)
requests>=2.31.0

# === ANÁLISE E RELATÓRIOS ===
# Processamento de texto
beautifulsoup4>=4.12.0

# Análise de user agents
user-agents>=2.2.0

# === MONITORAMENTO ===
# Verificação de sistema
psutil>=5.9.0

# === DESENVOLVIMENTO ===
# Formatação de código (opcional)
black>=23.0.0

# Linting (opcional)
flake8>=6.0.0

# === UTILITÁRIOS ===
# Manipulação de datas
python-dateutil>=2.8.2

# Expressões regulares avançadas
regex>=2023.0.0

# === VISUALIZAÇÃO (OPCIONAL) ===
# Gráficos para relatórios avançados
matplotlib>=3.7.0
seaborn>=0.12.0

# === SEGURANÇA ===
# Criptografia para tokens
cryptography>=41.0.0

# === NGROK (PARA TRACKING EXTERNO) ===
# Tunneling para servidor local
pyngrok>=6.0.0

# === COMPATIBILIDADE ===
# Suporte a timezone
pytz>=2023.3

# Encoding detection
chardet>=5.0.0

# === VERSÕES MÍNIMAS PYTHON ===
# Python >= 3.8 requerido